let an unpacked sdist run the test suite
The sdist shipped tests but not tasks/, so collection died importing
coverage_pragmas. Ship tasks/ and move the capability probes into a
module that does not pull in coverage, so a consumer building from
source can run the suite without installing the coverage extra.

Closes #684
