+++ tests/fork_helpers.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import os
 from typing import TYPE_CHECKING, Final, NoReturn, cast
 
-from coverage import Coverage, process_startup
+from capabilities import CAPABILITIES
 
 if TYPE_CHECKING:
     from collections.abc import Callable
@@ -39,20 +39,31 @@ def fork_process(child: Callable[[], NoReturn] | None = None) -> int:  # pragma:
 
 def exit_child(status: int) -> NoReturn:
     try:
-        if (coverage := Coverage.current()) is not None:
-            coverage.save()
+        _flush_coverage()
     finally:
         # A coverage write failure must not return the fork child to pytest.
         os._exit(status)
 
 
+def _flush_coverage() -> None:
+    # An unmeasured run has nothing to write, and importing coverage there would stop the suite from running at all.
+    if not CAPABILITIES["coverage"]:  # pragma: lacks coverage
+        return
+    from coverage import Coverage
+
+    if (coverage := Coverage.current()) is not None:
+        coverage.save()
+
+
 def _restart_coverage_after_fork() -> None:  # pragma: win32 no cover
+    from coverage import Coverage, process_startup
+
     if (coverage := Coverage.current()) is not None:  # pragma: win32 no cover
         coverage.stop()
     process_startup(force=True, slug="fork")  # pragma: no cover - coverage is stopped until this call returns
 
 
-if _REGISTER_AT_FORK is not None:  # pragma: win32 no cover
+if _REGISTER_AT_FORK is not None and CAPABILITIES["coverage"]:  # pragma: win32 no cover
     _REGISTER_AT_FORK(after_in_child=_restart_coverage_after_fork)
